Refactoring responses into new package
