+++ src/main/java/com/fairshare/Responses/CreateUserResponse.java
@@ -1,14 +1,14 @@
-package com.fairshare.Requests;
+package com.fairshare.Responses;
 
-public class CreateGroupResponse {
+public class CreateUserResponse {
     private String message;
     private Boolean success;
-    private Integer groupId;
+    private Integer userId;
 
-    public CreateGroupResponse(String message, Boolean success, Integer userId) {
+    public CreateUserResponse(String message, Boolean success, Integer userId) {
         this.message = message;
         this.success = success;
-        this.groupId = groupId;
+        this.userId = userId;
     }
 
     public String getMessage() {
@@ -27,11 +27,12 @@ public class CreateGroupResponse {
         this.success = success;
     }
 
-    public Integer getGroupId() {
-        return groupId;
+    public Integer getUserId() {
+        return userId;
     }
-    public void setGroupId(Integer groupId) {
-        this.groupId = groupId;
+
+    public void setUserId(Integer userId) {
+        this.userId = userId;
     }
 }
