+++ src/main/java/com/fairshare/Responses/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.fairshare.Requests;
+package com.fairshare.Responses;
 
 public class LoginResponse {
     private String message;
